create pages for graphs
