+++ apps/web/types/index.d.ts
@@ -94,6 +94,7 @@ export type DegreeSummary = {
 
 export type Pages = {
   Degrees: 'main' | 'add-new' | 'edit'
+  Graphs: 'main'
   Modules: 'main' | 'add-doing' | 'add-done'
   UserProfile:
     | 'public-profile'
